Integrate server runner with GLI app interface
Makes it possible to call server commands directly on the command line,
for example: 'mamp start mysql'. Sets up accepted servers and commands on the
bin file; that should be enough to setup the integration with the command line
interface, and for the underlying core to happily accept the commands, as they
are very similar in nature.